Hover over this for instructions
Notice the file names are in pairs.  One file name  in each pair contains the words "sans timestamps and linebreaks".  That is the file best suited for word and phrase searching because between every transcribed word is consistently a space and only one space.  The downside of this file is that timestamps had to be removed to produce that.

Hence the reason for the other file in a file name pair.  That file retains the YouTube timestamps, the purpose of which is to facilitate locating the time within a video where the search terms were found.  YouTube seems to place timestamps every 5 seconds or less.  Searches within this file are more difficult because there are often a variety of other characters between words.