How to contribute files to this project
This process uses Linux commands entered via the command line, so I expect you to have that level of skill.  The commands will be copy-pasted and begin with the filename= instruction.  Adjust the filename variable to reflect the file name you've made from YouTube transcripts.  TO MAKE THAT FILE FOLLOW THE PATTERN LAID OUT IN PREVIOUS TRANSCRIPT FILES.  IOW, some lines of intro are allowed ending with a blank line.  Each video transcript will be one to three lines for the video title, one line for whose YouTube account it was found under with the date they posted it, two lines for the two urls.  Two urls are the way YouTube has chosen - the normal is at https://www.youtube.com/watch?v= and the ->SHARE url which is at https://youtu.be/.  After those two lines is a blank line, then you'll paste in the transcript after that.  The transcript will look like every other line is a timestamp line and the lines of transcript are short.