Hover over this for instructions
On GitHub, you'll need to render these files "raw", implying that the first click you make on the file name will not be enough.  After that first click, a page will come up specific to the chosen file, but not showing it usefully.  You must click once again on a link that says "raw".  Look for it.  In raw rendering, word and phrase searches can be performed using the normal means of doing so by your particular browser.