Hover over this for instructions
On GitHub, if you see an option to render these files "raw", that rendering is the only means of displaying files as large in size as these transcripts are - you need to select "raw".  In raw rendering, word and phrase searches can be performed using the normal means of doing so by your particular browser.

Notice near the top where some files are listed.  The file names of Kat Kerr's presentations are in pairs.  One file name in each pair contains the words "sans timestamps and linebreaks".  That is the file best suited for word and phrase searching because between every transcribed word is consistently a space and only one space.  The downside of this file is that timestamps had to be removed to produce that.

Hence the reason for the other file in any file name pair.  That file retains the YouTube timestamps, the purpose of which is to facilitate locating the time within a video where the search terms were spoken.  YouTube seems to place timestamps every 5 seconds or less.  Searches within this file are more difficult because there are often a variety of other characters between words, but when you absolutely must locate the spoken words in a video, this file can help.